% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apricot_c.R
\name{apricot_c}
\alias{apricot_c}
\title{Apricot continuous quick-apply}
\usage{
apricot_c()
}
\value{
none
}
\description{
Applies the apricot theme and continuous color & fill scales to a ggplot
}
\examples{
apricot_c()
}
