% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apricot_d.R
\name{apricot_d}
\alias{apricot_d}
\title{Apricot discrete quick-apply}
\usage{
apricot_d()
}
\value{
none
}
\description{
Applies the apricot theme and discrete color & fill scales to a ggplot
}
\examples{
apricot_d()
}
