% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterfly_c.R
\name{butterfly_c}
\alias{butterfly_c}
\title{Butterfly continuous quick-apply}
\usage{
butterfly_c()
}
\value{
none
}
\description{
Applies the butterfly theme and continuous color & fill scales to a ggplot
}
\examples{
butterfly_c()
}
