% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterfly_d.R
\name{butterfly_d}
\alias{butterfly_d}
\title{Butterfly discrete quick-apply}
\usage{
butterfly_d()
}
\value{
none
}
\description{
Applies the butterfly theme and discrete color & fill scales to a ggplot
}
\examples{
butterfly_d()
}
