% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtea.R
\docType{package}
\name{ggtea}
\alias{ggtea}
\title{Palettes and Themes for 'ggplot2'}
\description{
A collection of palettes and themes for 'ggplot2', offering a light, pastel aesthetic. Syntax follows the 'viridis' package.
}
