% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herbal_c.R
\name{herbal_c}
\alias{herbal_c}
\title{Herbal continuous quick-apply}
\usage{
herbal_c()
}
\value{
none
}
\description{
Applies the herbal theme and continuous color & fill scales to a ggplot
}
\examples{
herbal_c()
}
