% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herbal_d.R
\name{herbal_d}
\alias{herbal_d}
\title{Herbal discrete quick-apply}
\usage{
herbal_d()
}
\value{
none
}
\description{
Applies the matcha theme and discrete color & fill scales to a ggplot
}
\examples{
herbal_d()
}
