% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcha_c.R
\name{matcha_c}
\alias{matcha_c}
\title{Matcha continuous quick-apply}
\usage{
matcha_c()
}
\value{
none
}
\description{
Applies the matcha theme and continuous color & fill scales to a ggplot
}
\examples{
matcha_c()
}
