% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcha_d.R
\name{matcha_d}
\alias{matcha_d}
\title{Matcha discrete quick-apply}
\usage{
matcha_d()
}
\value{
none
}
\description{
Applies the matcha theme and discrete color & fill scales to a ggplot
}
\examples{
matcha_d()
}
