% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_apricot_c.R
\name{scale_color_apricot_c}
\alias{scale_color_apricot_c}
\title{Apricot continuous color scale}
\usage{
scale_color_apricot_c()
}
\value{
none
}
\description{
Apply a continuous scale to a ggplot2 plot
}
\examples{
scale_color_apricot_c()
}
