% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_butterfly_d.R
\name{scale_color_butterfly_d}
\alias{scale_color_butterfly_d}
\title{Butterfly discrete color scale}
\usage{
scale_color_butterfly_d()
}
\value{
none
}
\description{
Apply an 8-color discrete scale to a ggplot2 plot
}
\examples{
scale_color_butterfly_d()
}
