% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_matcha_c.R
\name{scale_color_matcha_c}
\alias{scale_color_matcha_c}
\title{Matcha continuous color scale}
\usage{
scale_color_matcha_c()
}
\value{
none
}
\description{
Apply a continuous color scale to a ggplot2 plot
}
\examples{
scale_color_matcha_c()
}
