% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_herbal_c.R
\name{scale_fill_herbal_c}
\alias{scale_fill_herbal_c}
\title{Herbal continuous fill scale}
\usage{
scale_fill_herbal_c()
}
\value{
none
}
\description{
Apply a continuous fill scale to a ggplot2 plot
}
\examples{
scale_fill_herbal_c()
}
