% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_herbal_d.R
\name{scale_fill_herbal_d}
\alias{scale_fill_herbal_d}
\title{Herbal discrete fill scale}
\usage{
scale_fill_herbal_d()
}
\value{
none
}
\description{
Apply an 8-color discrete scale to a ggplot2 plot
}
\examples{
scale_fill_herbal_d()
}
