% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_butterfly.R
\name{theme_butterfly}
\alias{theme_butterfly}
\title{Butterfly theme for ggplot2}
\usage{
theme_butterfly()
}
\value{
none
}
\description{
Apply the apricot theme to a ggplot2 plot
}
\examples{
theme_butterfly()
}
